refactor: NoticeRegister 컴포넌트 생성 후 커스텀한 input, label 컴포넌트들 이동시킴
